Update Article and comment

+++ src/main/java/io/github/LucasZSGP/medium/domain/user/UserRepository.java
@@ -1,12 +1,11 @@
 /* (C)2024 */
 package io.github.LucasZSGP.medium.domain.user;
 
-import io.github.LucasZSGP.medium.domain.user.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     UserEntity findByEmail(String email);
 
     UserEntity findByUsername(String username);
